# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DFD API",
        "contact": {
            "name": "Seokho Song",
            "url": "http://github.com/devsdkGIN_MODE"
        },
        "license": {
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "devsdk.net/api/dfd",
    "basePath": "/",
    "paths": {
...
        "/v1/application/riot/access": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "1st party application token.\nPermission : **admin.token.create**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/admin"
                ],
                "summary": "Create new application token",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.RequestBodyToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/application/token": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "1st party application token.\nPermission : **admin.token.create**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/admin"
                ],
                "summary": "Create new application token",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docmodels.RequestBodyToken": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "riot-api-access"
                }
            }
        },
...
        "docmodels.ResponseBadRequest": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Bad request message"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "docmodels.ResponseInternalServerError": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
...
        "docmodels.ResponseNotFound": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Not found message"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "docmodels.ResponseSuccess": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "docmodels.ResponseUnauthorized": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Request Unauthorized message"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "token_expired": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}